git-svn-id: http://avisynth64.googlecode.com/svn/trunk@19 fa1118b8-3062-11df-954b-21e382534593

+++ distrib/include/SoftWire/Loader.cpp
@@ -161,7 +161,7 @@ namespace SoftWire
 			else if(encoding.hasImmediate() && encoding.relativeReference())
 			{
 				__int64 offset = encoding.getImmediate() - (__int64)currentCode - encoding.length(currentCode);
-				encoding.setCallOffset(offset);
+				encoding.setCallOffset((int)offset);
 			}
 
 			if(x64 && encoding.isRipRelative())
